requirements.txt
aiohttp==1.0.5
altgraph==0.16.1
async-timeout==3.0.0
asyncio==3.4.3
beautifulsoup4==4.6.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cycler==0.10.0
discord.py==0.16.12
future==0.17.1
html2text==2018.1.9
idna==2.7
kiwisolver==1.0.1
lxml==4.2.5
macholib==1.11
matplotlib==3.0.0
multidict==4.4.2
numpy==1.15.3
pandas==0.23.4
pefile==2018.8.8
Pillow==5.3.0
pycparser==2.19
PyInstaller==3.4
PyNaCl==1.0.1
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.7
pywin32-ctypes==0.2.0
requests==2.19.1
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
Unidecode==1.0.23
urllib3==1.23
utils==0.9.0
websockets==3.4
xmltodict==0.11.0
youtube-dl==2018.12.17
